+++ src/main/java/Team_Mute/back_end/domain/member/repository/AdminRegionRepository.java
@@ -12,5 +12,7 @@ public interface AdminRegionRepository extends JpaRepository<AdminRegion, Intege
 
 	Optional<AdminRegion> findByRegionName(String regionName);
 
+	Optional<AdminRegion> findByRegionId(Integer regionId);
+
 	boolean existsByRegionName(String regionName);
 }
