+++ src/main/java/Team_Mute/back_end/domain/space_admin/controller/SpaceController.java
@@ -1,5 +1,23 @@
 package Team_Mute.back_end.domain.space_admin.controller;
 
+import java.util.List;
+import java.util.Map;
+import java.util.NoSuchElementException;
+
+import org.springframework.http.MediaType;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.PutMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestPart;
+import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.multipart.MultipartFile;
+
+import com.fasterxml.jackson.databind.ObjectMapper;
+
 import Team_Mute.back_end.domain.space_admin.dto.CategoryListItem;
 import Team_Mute.back_end.domain.space_admin.dto.DeleteSpaceResponse;
 import Team_Mute.back_end.domain.space_admin.dto.LocationListItem;
@@ -21,22 +39,6 @@ import io.swagger.v3.oas.annotations.parameters.RequestBody;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
-
-import java.util.List;
-import java.util.Map;
-import java.util.NoSuchElementException;
-import org.springframework.http.MediaType;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.DeleteMapping;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.PutMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestPart;
-import org.springframework.web.bind.annotation.RestController;
-import org.springframework.web.multipart.MultipartFile;
-import com.fasterxml.jackson.databind.ObjectMapper;
 
 @Tag(name = "공간 관리 API", description = "관리자 공간 관리 관련 API 명세")
 @RestController
@@ -120,7 +122,8 @@ public class SpaceController {
 			// String 데이터를 SpaceCreateRequest 객체로 수동 변환
 			SpaceCreateRequest request = objectMapper.readValue(spaceJson, SpaceCreateRequest.class);
 
-			boolean noUsableFiles = (images == null || images.isEmpty()) || images.stream().allMatch(f -> f == null || f.isEmpty());
+			boolean noUsableFiles =
+				(images == null || images.isEmpty()) || images.stream().allMatch(f -> f == null || f.isEmpty());
 			if (noUsableFiles) {
 				return ResponseEntity.badRequest().body("이미지는 최소 1장은 필요합니다.");
 			}
@@ -166,7 +169,8 @@ public class SpaceController {
 			SpaceCreateRequest request = objectMapper.readValue(spaceJson, SpaceCreateRequest.class);
 
 			// 이미지가 없을 경우 예외 처리
-			boolean noUsableFiles = (images == null || images.isEmpty()) || images.stream().allMatch(f -> f == null || f.isEmpty());
+			boolean noUsableFiles =
+				(images == null || images.isEmpty()) || images.stream().allMatch(f -> f == null || f.isEmpty());
 
 			if (noUsableFiles) {
 				return ResponseEntity.badRequest().body("이미지는 최소 1장은 필요합니다.");
@@ -210,4 +214,3 @@ public class SpaceController {
 		return ResponseEntity.ok(result);
 	}
 }
-
